+++ esa_climate_toolbox/ds/dataframeaccess.py
@@ -19,20 +19,15 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from datetime import datetime
-from dateutil.relativedelta import relativedelta
 from typing import Any
-from typing import List
 from typing import Mapping
 from typing import Tuple
-from typing import Union
 import geopandas as gpd
 import pandas as pd
 import shapely
 
-from ..constants import MONTHS
-from .constants import TIMESTAMP_FORMAT
 from esa_climate_toolbox.ds.ccicdc import CciCdc
+from esa_climate_toolbox.ds.timerangegetter import TimeRangeGetter
 from esa_climate_toolbox.core.types import GeoDataFrame
 
 
@@ -56,101 +51,8 @@ class DataFrameAccessor:
             self._spatial_subset_area = shapely.Polygon(coords)
         # we use a time chunking of 1 as we do not know the actual content of a nc file
         self._time_chunking = 1
-        self._time_ranges = self._get_time_ranges(self._df_id, gdf_params)
-
-    def _get_time_ranges(self, dataset_id: str,
-                         params: Mapping[str, Any]) -> List[Tuple]:
-        start_time, end_time, iso_start_time, iso_end_time = \
-            self._extract_time_range_as_datetime(
-                params.get('time_range',
-                           self.get_default_time_range(dataset_id)))
-        time_period = dataset_id.split('.')[2]
-        if time_period == 'day':
-            start_time = datetime(year=start_time.year, month=start_time.month,
-                                  day=start_time.day)
-            end_time = datetime(year=end_time.year, month=end_time.month,
-                                day=end_time.day,
-                                hour=23, minute=59, second=59)
-            delta = relativedelta(days=1)
-        elif time_period == 'month' or time_period == 'mon':
-            start_time = datetime(year=start_time.year, month=start_time.month,
-                                  day=1)
-            end_time = datetime(year=end_time.year, month=end_time.month, day=1)
-            delta = relativedelta(months=1)
-            end_time += delta
-        elif time_period == 'year' or time_period == 'yr':
-            start_time = datetime(year=start_time.year, month=1, day=1)
-            end_time = datetime(year=end_time.year, month=12, day=31)
-            delta = relativedelta(years=1)
-        elif time_period == 'climatology':
-            return [(i + 1, i + 1) for i, month in enumerate(MONTHS)]
-        else:
-            end_time = end_time.replace(hour=23, minute=59, second=59)
-            end_time_str = datetime.strftime(end_time, TIMESTAMP_FORMAT)
-            iso_end_time = self._extract_time_as_string(end_time_str)
-            request_time_ranges = self._cci_cdc.get_time_ranges_from_data(
-                dataset_id, iso_start_time, iso_end_time)
-            return request_time_ranges
-        request_time_ranges = []
-        this = start_time
-        while this < end_time:
-            after = this + delta
-            pd_this = pd.Timestamp(datetime.strftime(this, TIMESTAMP_FORMAT))
-            pd_next = pd.Timestamp(datetime.strftime(after, TIMESTAMP_FORMAT))
-            request_time_ranges.append((pd_this, pd_next))
-            this = after
-        return request_time_ranges
-
-    def _extract_time_range_as_datetime(
-            self, time_range: Union[Tuple, List]
-    ) -> (datetime, datetime, str, str):
-        iso_start_time, iso_end_time = self._extract_time_range_as_strings(
-            time_range)
-        start_time = datetime.strptime(iso_start_time, TIMESTAMP_FORMAT)
-        end_time = datetime.strptime(iso_end_time, TIMESTAMP_FORMAT)
-        return start_time, end_time, iso_start_time, iso_end_time
-
-    @classmethod
-    def _extract_time_range_as_strings(
-            cls, time_range: Union[Tuple, List]
-    ) -> (str, str):
-        if isinstance(time_range, tuple):
-            time_start, time_end = time_range
-        else:
-            time_start = time_range[0]
-            time_end = time_range[1]
-        return \
-            cls._extract_time_as_string(time_start),\
-            cls._extract_time_as_string(time_end)
-
-    @classmethod
-    def _extract_time_as_string(
-            cls, time_value: Union[pd.Timestamp, str]
-    ) -> str:
-        if isinstance(time_value, str):
-            time_value = pd.to_datetime(time_value, utc=True)
-        return time_value.tz_localize(None).isoformat()
-
-    def get_default_time_range(self, ds_id: str):
-        temporal_start = self._metadata.get('temporal_coverage_start', None)
-        temporal_end = self._metadata.get('temporal_coverage_end', None)
-        if not temporal_start or not temporal_end:
-            time_ranges = self._cci_cdc.get_time_ranges_from_data(ds_id)
-            if not temporal_start:
-                if len(time_ranges) == 0:
-                    raise ValueError(
-                        "Could not determine temporal start of dataset. "
-                        "Please use 'time_range' parameter."
-                    )
-                temporal_start = time_ranges[0][0]
-            if not temporal_end:
-                if len(time_ranges) == 0:
-                    raise ValueError(
-                        "Could not determine temporal end of dataset. "
-                        "Please use 'time_range' parameter."
-                    )
-                temporal_end = time_ranges[-1][1]
-        return temporal_start, temporal_end
+        tr = TimeRangeGetter(self._cci_cdc, self._metadata)
+        self._time_ranges = tr.get_time_ranges(self._df_id, gdf_params)
 
     def request_time_range(self, time_index: int) -> Tuple:
         start_index = time_index * self._time_chunking
